2.0.0 - Working sample (#7)
* Foto met dora als test document

* Auto delete file, and also much better README.md

* rm use of appwrite client

* readme update

* New test documents

* Mime type handler, which can be useful for tracking what mime type
everything is

* Avoidance of regex

* Simple login page made

* JobTask tracking, using react as frontend so that we have more
flexiblity

* Implementation of /api/job_status/<str:job_task_id>

* Email validation

* My .gitignore was disrupting me

* Add of bootstrap

* React frontend (#4)

* JobTask tracking, using react as frontend so that we have more
flexiblity

* Implementation of /api/job_status/<str:job_task_id>

* Email validation

* My .gitignore was disrupting me

* Add of bootstrap

* Sample toml file

* goback.toml replacement (we don't want .env files!!!)

* Local development .gitignore

* Rename symbols

* Dynamic Credentials holder

* Start of fastapi

* Working fastapi-example

* Renaming

* Session mistake

* Working example, plus also credentials validation for appwrite

In this commit, I have made a checker that can check if the appwrite
credentials are partially correct or not, this will kill/ reduce the
amount of hassle the user can have later on when something doesn't work

* Use of rich module (later on we'll use logging)

* Removal of old react project

We are going to make a new one

* Home page finished

Home page is finished, now we're going to the forms (login & signup
pages)

* Add min length of user

Much better, otherwise you'll have usernames that are 250+ characters :P

* Signup form finish

In this commit, I have finished creating te signup form in Signup.tsx,
we're going to focus now on authentication

* Add of protected routes

Protected routes means that the user must be authorized/authenticated to
access the page, which is needed.
We also extended the ``Login.tsx`` functionality

* Favicon.ico

* Cors

* Protected route has been updated, and much more

Login is working now, and we also have toasts now to notify the user
authentication works too now, except /signup which we're going to
implement in the next commit

* More stuff

In this commit, I have made some pages:
> CreateBackupPage.tsx -> The page to create a backup/snapshot

Some things has been changed, such as models.py for validation

* URL validation

* New route,

A new route has been createn, this is, /job?job_id=job_id, which can be
used to view a job's status

* Stop requesting when success of job

* List as status

In this commit, status has been replaced/renewed with status-messages,
which holds, a serialized python3 list of models.Status objects.

* Types

* Dumb linting

* Link to show the result

* Linting

* Update requirements.txt

* Serving react builded files

In this commit, I managed to make it work that fastapi, the webserver,
can serve react frontend files, but the files should still be builded

* Removal of function and validation of appwrite URL endpoint

* Requirements.txt

* Navbar created

* Whoops forgot this:

* New branch for docker

* update requirements.txt

* Dockerilfe

* Revert "update requirements.txt"

This reverts commit 7dcb30336f33037afdac99661ca181fc23893ba3.

* rm docker

* update

* Jobs route

In this commit, new route

* Update job list

In this commit, i have improved the /jobs route

* Order of routes

Small change, but here I have changed the order of the routes,
especially putted logout as last

* Dockerfile

* Propper error handling when registering

Small fix, which is checking if status code is 400 when registering an
user.

* Conditional routes

if ./goback-frontend/dist/assets is missing then it should just not
serve some routes

* Working compose.yaml